+++ src/LanManager.Maui.Crew/Views/CheckInPage.xaml.cs
@@ -1,0 +1,13 @@
+using LanManager.Maui.Crew.ViewModels;
+
+namespace LanManager.Maui.Crew.Views;
+
+public partial class CheckInPage : ContentPage
+{
+    public CheckInPage(CheckInViewModel viewModel)
+    {
+        InitializeComponent();
+        BindingContext = viewModel;
+    }
+}
+